Added "no_sunlight"
"no_sunlight" has been added as an option for window photolysis allowing experiments with no solar photolysis to be modelled. The "off" setting for indoor lighting was already in the model but this has now been added to the manual alongside "no_sunlight". It is now possible to run the model with no photolysis indoors.