Updated CO outdoor concentration
CO outdoors varies massively depending on location. The previous value was an Iranian value as published material was limited. A new value, the medial value of the annual mean from stations across the UK in 2016, has been used instead from the European Air Database. Citation in the GMD paper for INCHEM-Py 1.2