Timed emissions update
Previously timed emissions were set by including a value in the master array for each species of "species_timed". This would be the rate of increase for that species and the value would be set to 0 if the current simulation time was between the emission start and end time.

This update removes that method. The timed emissions are now put in as reactions into the reactions set as "species_timed_1 * rate = species", where species_timed_1 is set to 0 or 1 depending on the current time. This means that strings can be put in as rates into the timed emissions (such as multiplying by the ACRate to account for measured losses).