Updates to particle calculations
Particle calculations have been lower than expected, or previously delivered. This was determined to be due to the reactions not including HOMs (highly oxygenated molecules) which have been included in previous INDCM model simulations.

The fix for this includes the splitting of seed particles into those created by chemistry and those sourced from outdoors. We have also changed the scaling factor in the partitioning of gas off particles. This has been changed to its old value of 120 from Carslaw 2012 but will be removed in future updates as we reintroduce HOMs back into the model.

This resolves #17, future updates will improve particle representation but we are happy that the results given now are representative.