Events and volumes
Includes two updates, resolves #18 and also an addition of an events processing system in the integration.

Resolution of #18 involves the splitting of the surface to volume ratios into the individual input of volume and surface areas. This is easier for the user to understand and makes the implementation less complex. It also allows for breath emissions to be adjusted to the room volume.

Events are now also included. Events are points in the simulation where the integration is determined to break continuity. These have been manually set as sunrise, sunset and any timed emissions. At these points the integrator restarts and has led to a not-insignificant improvement in model run time.

It does mean that there is increased importance in making sure the timed emissions are on the integrator external time steps.